chore: move person schema to examples
