+++ packages/storybook-angular/src/renderers/advanced.stories.ts
@@ -1,8 +1,8 @@
 import { storiesOf } from '@storybook/angular';
 import { action } from '@storybook/addon-actions';
 import { IFormSubmission } from '@de-re-crud/angular';
+import schema from '../../../../examples/schemas/person.json';
 import { CustomRendererModule } from './app/custom-renderer.module';
-import schema from './person-schema.json';
 
 storiesOf('Renderers/Advanced Examples', module).add(
   'recursive (collapsible)',
