--- packages/storybook-ui/src/renderers/person-schema.json
@@ -1,84 +1,0 @@
-{
-  "structs": [
-    {
-      "name": "person",
-      "label": "Person",
-      "collectionLabel": "People",
-      "blocks": [
-        {
-          "name": "default",
-          "references": [
-            "ssn",
-            {
-              "block": "name"
-            },
-            {
-              "field": "children",
-              "hints": {
-                "block": "name",
-                "layout": "table"
-              }
-            }
-          ]
-        },
-        {
-          "name": "name",
-          "references": ["firstName", "middleName", "lastName"],
-          "hints": {
-            "layout": "horizontal"
-          }
-        }
-      ],
-      "fields": [
-        {
-          "name": "ssn",
-          "label": "SSN",
-          "type": "text",
-          "keyField": true,
-          "required": true
-        },
-        {
-          "name": "firstName",
-          "label": "First Name",
-          "type": "text",
-          "required": true,
-          "hints": {
-            "width": 5
-          }
-        },
-        {
-          "name": "middleName",
-          "label": "M",
-          "placeholder": "Middle Name",
-          "type": "text",
-          "hints": {
-            "width": 2
-          }
-        },
-        {
-          "name": "lastName",
-          "label": "Last Name",
-          "type": "text",
-          "required": true,
-          "hints": {
-            "width": 5
-          }
-        },
-        {
-          "name": "children",
-          "label": "Children",
-          "type": "linkedStruct",
-          "reference": {
-            "struct": "person"
-          },
-          "hints": {
-            "layout": "table",
-            "custom": {
-              "childKey": "children"
-            }
-          }
-        }
-      ]
-    }
-  ]
-}
